[JavaCallableWrappers] Begin modularizing jcw process.
