+++ src/Java.Interop.Tools.JavaCallableWrappers/Java.Interop.Tools.JavaCallableWrappers/JavaCallableMethodClassifier.cs
@@ -1,0 +1,9 @@
+﻿using Mono.Cecil;
+
+namespace Java.Interop.Tools.JavaCallableWrappers
+{
+	public abstract class JavaCallableMethodClassifier
+	{
+		public abstract bool ShouldBeDynamicallyRegistered (TypeDefinition topType, MethodDefinition registeredMethod, MethodDefinition implementedMethod, CustomAttribute? registerAttribute);
+	}
+}
