+++ src/Java.Interop.Tools.JavaCallableWrappers/Java.Interop.Tools.JavaCallableWrappers/JavaCallableWrapperGenerator.cs
@@ -20,49 +20,14 @@ using static Java.Interop.Tools.TypeNameMappings.JavaNativeTypeManager;
 
 namespace Java.Interop.Tools.JavaCallableWrappers {
 
-	public enum JavaPeerStyle {
-		XAJavaInterop1,
-		JavaInterop1,
-	}
-
-	public abstract class JavaCallableMethodClassifier
-	{
-		public abstract bool ShouldBeDynamicallyRegistered (TypeDefinition topType, MethodDefinition registeredMethod, MethodDefinition implementedMethod, CustomAttribute? registerAttribute);
-	}
-
-	 public class JavaCallableWrapperGenerator {
-
-		class JavaFieldInfo {
-			public JavaFieldInfo (MethodDefinition method, string fieldName, IMetadataResolver resolver)
-			{
-				this.FieldName = fieldName;
-				InitializerName = method.Name;
-				TypeName = JavaNativeTypeManager.ReturnTypeFromSignature (GetJniSignature (method, resolver))?.Type
-					?? throw new ArgumentException ($"Could not get JNI signature for method `{method.Name}`", nameof (method));
-				IsStatic = method.IsStatic;
-				Access = method.Attributes & MethodAttributes.MemberAccessMask;
-				Annotations = GetAnnotationsString ("\t", method.CustomAttributes, resolver);
-			}
-
-			public MethodAttributes Access { get; private set; }
-			public bool IsStatic { get; private set; }
-			public string TypeName { get; private set; }
-			public string FieldName { get; private set; }
-			public string InitializerName { get; private set; }
-			public string Annotations { get; private set; }
-
-			public string GetJavaAccess ()
-			{
-				return JavaCallableWrapperGenerator.GetJavaAccess (Access);
-			}
-		}
+	public partial class JavaCallableWrapperGenerator {
 
 		Action<string, object[]> log;
-		string name;
-		string package;
-		TypeDefinition type;
+		internal string name;
+		internal string package;
+		internal TypeDefinition type;
 		List<JavaFieldInfo> exported_fields = new List<JavaFieldInfo> ();
-		List<Signature> methods = new List<Signature> ();
+		internal List<Signature> methods = new List<Signature> ();
 		List<Signature> ctors   = new List<Signature> ();
 		List<JavaCallableWrapperGenerator>? children;
 
@@ -575,91 +540,35 @@ namespace Java.Interop.Tools.JavaCallableWrappers {
 			return sb.ToString ();
 		}
 
-		// example of java target to generate for a type
-		//
-		// package mono.droid;
-		//
-		// import android.app.Activity;
-		// import android.os.Bundle;
-		//
-		// public class MonoActivity extends android.app.Activity
-		// {
-		// 	  static final String __md_methods;
-		// 	  static {
-		// 	    __md_methods =
-		// 	      "n_OnCreate:(Landroid/os/Bundle;)V:GetOnCreate_Landroid_os_Bundle_Handler\n" +
-		// 	      "";
-		// 	    mono.android.Runtime.register ("Mono.Droid.MonoActivity, AssemblyName", MonoActivity.class, __md_methods);
-		// 	  }
-		//
-		//    public void onCreate(android.os.Bundle savedInstanceState)
-		//    {
-		//      n_onCreate (savedInstanceState);
-		//    }
-		//
-		//    private native void n_onCreate (android.os.Bundle bundle);
-		// }
+		public void Generate (string outputPath)
+		{
+			var jcw_writer = CreateWriter ();
+			jcw_writer.Generate (outputPath);
+		}
 
 		public void Generate (TextWriter writer)
 		{
-			if (!string.IsNullOrEmpty (package)) {
-				writer.WriteLine ("package " + package + ";");
-				writer.WriteLine ();
-			}
-
-			GenerateHeader (writer);
-
-			bool needCtor = false;
-			if (HasDynamicallyRegisteredMethods) {
-				needCtor = true;
-				writer.WriteLine ("/** @hide */");
-				writer.WriteLine ("\tpublic static final String __md_methods;");
-			}
-
-			if (children != null) {
-				for (int i = 0; i < children.Count; i++) {
-					if (!children[i].HasDynamicallyRegisteredMethods) {
-						continue;
-					}
-					needCtor = true;
-					writer.Write ("\tstatic final String __md_");
-					writer.Write (i + 1);
-					writer.WriteLine ("_methods;");
-				}
-			}
-
-			if (needCtor) {
-				writer.WriteLine ("\tstatic {");
-
-				if (HasDynamicallyRegisteredMethods) {
-					GenerateRegisterType (writer, this, "__md_methods");
-				}
-
-				if (children != null) {
-					for (int i = 0; i < children.Count; ++i) {
-						GenerateRegisterType (writer, children [i], $"__md_{i + 1}_methods");
-					}
-				}
-				writer.WriteLine ("\t}");
-			}
-
-			GenerateBody (writer);
-
-			if (children != null)
-				foreach (JavaCallableWrapperGenerator child in children) {
-					child.GenerateHeader (writer);
-					child.GenerateBody (writer);
-					child.GenerateFooter (writer);
-				}
-
-			GenerateFooter (writer);
+			var jcw_writer = CreateWriter ();
+			jcw_writer.Generate (writer);
 		}
 
-		public void Generate (string outputPath)
+		internal JavaCallableWrapperWriter CreateWriter ()
 		{
-			using (StreamWriter sw = OpenStream (outputPath)) {
-				Generate (sw);
-			}
+			return new JavaCallableWrapperWriter (
+				name: name,
+				package: package,
+				type: type,
+				hasDynamicallyRegisteredMethods: HasDynamicallyRegisteredMethods,
+				children: children,
+				cache: cache,
+				applicationJavaClass: ApplicationJavaClass,
+				codeGenerationTarget: CodeGenerationTarget,
+				generateOnCreateOverrides: GenerateOnCreateOverrides,
+				monoRuntimeInitialization: MonoRuntimeInitialization,
+				exported_fields: exported_fields,
+				methods: methods,
+				ctors: ctors,
+				generator: this);
 		}
 
 		static string GetAnnotationsString (string indent, IEnumerable<CustomAttribute> atts, IMetadataResolver resolver)
@@ -669,7 +578,7 @@ namespace Java.Interop.Tools.JavaCallableWrappers {
 			return sw.ToString ();
 		}
 
-		static void WriteAnnotations (string indent, TextWriter sw, IEnumerable<CustomAttribute> atts, IMetadataResolver resolver)
+		internal static void WriteAnnotations (string indent, TextWriter sw, IEnumerable<CustomAttribute> atts, IMetadataResolver resolver)
 		{
 			foreach (var ca in atts) {
 				var catype = resolver.Resolve (ca.AttributeType);
@@ -698,6 +607,20 @@ namespace Java.Interop.Tools.JavaCallableWrappers {
 			}
 		}
 
+		static string GetJavaAccess (MethodAttributes access)
+		{
+			switch (access) {
+				case MethodAttributes.Public:
+					return "public";
+				case MethodAttributes.FamORAssem:
+					return "protected";
+				case MethodAttributes.Family:
+					return "protected";
+				default:
+					return "private";
+			}
+		}
+
 		// FIXME: this is hacky. Is there any existing code for value to source conversion?
 		static string ManagedValueToJavaSource (object value)
 		{
@@ -709,468 +632,6 @@ namespace Java.Interop.Tools.JavaCallableWrappers {
 				return ((bool) value) ? "true" : "false";
 			else
 				return value.ToString ();
-		}
-
-		void GenerateHeader (TextWriter sw)
-		{
-			sw.WriteLine ();
-
-			// class annotations.
-			WriteAnnotations ("", sw, type.CustomAttributes, cache);
-
-			sw.WriteLine ("public " + (type.IsAbstract ? "abstract " : "") + "class " + name);
-
-			string extendsType = GetJavaTypeName (type.BaseType, cache);
-			if (extendsType == "android.app.Application" && ApplicationJavaClass != null && !string.IsNullOrEmpty (ApplicationJavaClass))
-				extendsType = ApplicationJavaClass;
-			sw.WriteLine ("\textends " + extendsType);
-			sw.WriteLine ("\timplements");
-			sw.Write ("\t\t");
-			switch (CodeGenerationTarget) {
-				case JavaPeerStyle.JavaInterop1:
-					sw.Write ("net.dot.jni.GCUserPeerable");
-					break;
-				default:
-					sw.Write ("mono.android.IGCUserPeer");
-					break;
-			}
-			foreach (var ifaceInfo in type.Interfaces) {
-				var iface = cache.Resolve(ifaceInfo.InterfaceType);
-				if (!GetTypeRegistrationAttributes (iface).Any ())
-					continue;
-				sw.WriteLine (",");
-				sw.Write ("\t\t");
-				sw.Write (GetJavaTypeName (iface, cache));
-			}
-			sw.WriteLine ();
-			sw.WriteLine ("{");
-		}
-
-		void GenerateBody (TextWriter sw)
-		{
-			foreach (Signature ctor in ctors) {
-				if (string.IsNullOrEmpty (ctor.Params) && JavaNativeTypeManager.IsApplication (type, cache))
-					continue;
-				GenerateConstructor (ctor, sw);
-			}
-
-			GenerateApplicationConstructor (sw);
-
-			foreach (JavaFieldInfo field in exported_fields)
-				GenerateExportedField (field, sw);
-
-			foreach (Signature method in methods)
-				GenerateMethod (method, sw);
-
-			if (GenerateOnCreateOverrides && JavaNativeTypeManager.IsApplication (type, cache) && !methods.Any (m => m.Name == "onCreate"))
-				WriteApplicationOnCreate (sw, w => {
-						w.Write ("\t\tmono.android.Runtime.register (\"");
-						w.Write (type.GetPartialAssemblyQualifiedName (cache));
-						w.Write ("\", ");
-						w.Write (name);
-						w.WriteLine (".class, __md_methods);");
-						w.WriteLine ("\t\tsuper.onCreate ();");
-				});
-			if (GenerateOnCreateOverrides && JavaNativeTypeManager.IsInstrumentation (type, cache) && !methods.Any (m => m.Name == "onCreate"))
-				WriteInstrumentationOnCreate (sw, w => {
-						w.Write ("\t\tmono.android.Runtime.register (\"");
-						w.Write (type.GetPartialAssemblyQualifiedName (cache));
-						w.Write ("\", ");
-						w.Write (name);
-						w.WriteLine (".class, __md_methods);");
-						w.WriteLine ("\t\tsuper.onCreate (arguments);");
-				});
-
-			string addRef       = "monodroidAddReference";
-			string clearRefs    = "monodroidClearReferences";
-			if (CodeGenerationTarget == JavaPeerStyle.JavaInterop1) {
-				addRef      = "jiAddManagedReference";
-				clearRefs   = "jiClearManagedReferences";
-			}
-
-			sw.WriteLine ();
-			sw.WriteLine ("\tprivate java.util.ArrayList refList;");
-			sw.WriteLine ($"\tpublic void {addRef} (java.lang.Object obj)");
-			sw.WriteLine ("\t{");
-			sw.WriteLine ("\t\tif (refList == null)");
-			sw.WriteLine ("\t\t\trefList = new java.util.ArrayList ();");
-			sw.WriteLine ("\t\trefList.add (obj);");
-			sw.WriteLine ("\t}");
-			sw.WriteLine ();
-			sw.WriteLine ($"\tpublic void {clearRefs} ()");
-			sw.WriteLine ("\t{");
-			sw.WriteLine ("\t\tif (refList != null)");
-			sw.WriteLine ("\t\t\trefList.clear ();");
-			sw.WriteLine ("\t}");
-		}
-
-		void GenerateRegisterType (TextWriter sw, JavaCallableWrapperGenerator self, string field)
-		{
-			if (!self.HasDynamicallyRegisteredMethods) {
-				return;
-			}
-
-			sw.Write ("\t\t");
-			sw.Write (field);
-			sw.WriteLine (" = ");
-			string managedTypeName = self.type.GetPartialAssemblyQualifiedName (cache);
-			string javaTypeName = $"{package}.{name}";
-
-			foreach (Signature method in self.methods) {
-				if (method.IsDynamicallyRegistered) {
-					sw.Write ("\t\t\t\"", method.Method);
-					sw.Write (method.Method);
-					sw.WriteLine ("\\n\" +");
-				}
-			}
-			sw.WriteLine ("\t\t\t\"\";");
-			if (CannotRegisterInStaticConstructor (self.type))
-				return;
-			sw.Write ("\t\t");
-			switch (CodeGenerationTarget) {
-				case JavaPeerStyle.JavaInterop1:
-					sw.Write ("net.dot.jni.ManagedPeer.registerNativeMembers (");
-					sw.Write (self.name);
-					sw.Write (".class, ");
-					sw.Write (field);
-					sw.WriteLine (");");
-					break;
-				default:
-					sw.Write ("mono.android.Runtime.register (\"");
-					sw.Write (managedTypeName);
-					sw.Write ("\", ");
-					sw.Write (self.name);
-					sw.Write (".class, ");
-					sw.Write (field);
-					sw.WriteLine (");");
-					break;
-			}
-		}
-
-		void GenerateFooter (TextWriter sw)
-		{
-			sw.WriteLine ("}");
-		}
-
-		static string GetJavaAccess (MethodAttributes access)
-		{
-			switch (access) {
-			case MethodAttributes.Public:
-				return "public";
-			case MethodAttributes.FamORAssem:
-				return "protected";
-			case MethodAttributes.Family:
-				return "protected";
-			default:
-				return "private";
-			}
-		}
-
-		static string GetJavaTypeName (TypeReference r, IMetadataResolver cache)
-		{
-			TypeDefinition d = cache.Resolve (r);
-			string? jniName = JavaNativeTypeManager.ToJniName (d, cache);
-			if (jniName == null) {
-				Diagnostic.Error (4201, Localization.Resources.JavaCallableWrappers_XA4201, r.FullName);
-				throw new InvalidOperationException ("--nrt:jniName-- Should not be reached");
-			}
-			return jniName.Replace ('/', '.').Replace ('$', '.');
-		}
-
-		bool CannotRegisterInStaticConstructor (TypeDefinition type)
-		{
-			return JavaNativeTypeManager.IsApplication (type, cache) || JavaNativeTypeManager.IsInstrumentation (type, cache);
-		}
-
-		class Signature {
-
-			public Signature (MethodDefinition method, RegisterAttribute register, IMetadataResolver cache, bool shouldBeDynamicallyRegistered = true) : this (method, register, null, null, cache, shouldBeDynamicallyRegistered) {}
-
-			public Signature (MethodDefinition method, RegisterAttribute register, string? managedParameters, string? outerType, IMetadataResolver cache, bool shouldBeDynamicallyRegistered = true)
-				: this (register.Name, register.Signature, register.Connector, managedParameters, outerType, null)
-			{
-				Annotations = JavaCallableWrapperGenerator.GetAnnotationsString ("\t", method.CustomAttributes, cache);
-				IsDynamicallyRegistered = shouldBeDynamicallyRegistered;
-			}
-
-			public Signature (MethodDefinition method, ExportAttribute export, string? managedParameters, IMetadataResolver cache)
-				: this (method.Name, GetJniSignature (method, cache), "__export__", null, null, export.SuperArgumentsString)
-			{
-				IsExport = true;
-				IsStatic = method.IsStatic;
-				JavaAccess = JavaCallableWrapperGenerator.GetJavaAccess (method.Attributes & MethodAttributes.MemberAccessMask);
-				ThrownTypeNames = export.ThrownNames;
-				JavaNameOverride = export.Name;
-				ManagedParameters = managedParameters;
-				Annotations = JavaCallableWrapperGenerator.GetAnnotationsString ("\t", method.CustomAttributes, cache);
-			}
-
-			public Signature (MethodDefinition method, ExportFieldAttribute exportField, IMetadataResolver cache)
-				: this (method.Name, GetJniSignature (method, cache), "__export__", null, null, null)
-			{
-				if (method.HasParameters)
-					Diagnostic.Error (4205, JavaCallableWrapperGenerator.LookupSource (method), Localization.Resources.JavaCallableWrappers_XA4205);
-				if (method.ReturnType.MetadataType == MetadataType.Void)
-					Diagnostic.Error (4208, JavaCallableWrapperGenerator.LookupSource (method), Localization.Resources.JavaCallableWrappers_XA4208);
-				IsExport = true;
-				IsStatic = method.IsStatic;
-				JavaAccess = JavaCallableWrapperGenerator.GetJavaAccess (method.Attributes & MethodAttributes.MemberAccessMask);
-
-				// annotations are processed within JavaFieldInfo, not the initializer method. So we don't generate them here.
-			}
-
-			public Signature (string name, string? signature, string? connector, string? managedParameters, string? outerType, string? superCall)
-			{
-				ManagedParameters = managedParameters;
-				JniSignature      = signature ?? throw new ArgumentNullException ("`connector` cannot be null.", nameof (connector));
-				Method    = "n_" + name + ":" + JniSignature + ":" + connector;
-				Name      = name;
-
-				var jnisig = JniSignature;
-				int closer = jnisig.IndexOf (')');
-				string ret = jnisig.Substring (closer + 1);
-				retval = JavaNativeTypeManager.Parse (ret)?.Type;
-				string jniparms = jnisig.Substring (1, closer - 1);
-				if (string.IsNullOrEmpty (jniparms) && string.IsNullOrEmpty (superCall))
-					return;
-				var parms = new StringBuilder ();
-				var scall = new StringBuilder ();
-				var acall = new StringBuilder ();
-				bool first = true;
-				int i = 0;
-				foreach (JniTypeName jti in JavaNativeTypeManager.FromSignature (jniparms)) {
-					if (outerType != null) {
-						acall.Append (outerType).Append (".this");
-						outerType = null;
-						continue;
-					}
-					string? parmType = jti.Type;
-					if (!first) {
-						parms.Append (", ");
-						scall.Append (", ");
-						acall.Append (", ");
-					}
-					first = false;
-					parms.Append (parmType).Append (" p").Append (i);
-					scall.Append ("p").Append (i);
-					acall.Append ("p").Append (i);
-					++i;
-				}
-				this.parms = parms.ToString ();
-				this.call  = superCall != null ? superCall : scall.ToString ();
-				this.ActivateCall = acall.ToString ();
-			}
-
-			string? call;
-			public string? SuperCall {
-				get { return call; }
-			}
-
-			public string? ActivateCall {get; private set;}
-
-			public readonly string Name;
-			public readonly string? JavaNameOverride;
-			public string JavaName {
-				get { return JavaNameOverride ?? Name; }
-			}
-
-			string? parms;
-			public string? Params {
-				get { return parms; }
-			}
-
-			string? retval;
-			public string? Retval {
-				get { return retval; }
-			}
-
-			public string? ThrowsDeclaration {
-				get { return ThrownTypeNames?.Length > 0 ? " throws " + String.Join (", ", ThrownTypeNames) : null; }
-			}
-
-			public readonly string? JavaAccess;
-			public readonly string? ManagedParameters;
-			public readonly string JniSignature;
-			public readonly string Method;
-			public readonly bool IsExport;
-			public readonly bool IsStatic;
-			public readonly bool IsDynamicallyRegistered = true;
-			public readonly string []? ThrownTypeNames;
-			public readonly string? Annotations;
-		}
-
-		void GenerateConstructor (Signature ctor, TextWriter sw)
-		{
-			// TODO:  we only generate constructors so that Android types w/ no
-			//        default constructor can be subclasses by our generated code.
-			//
-			//        This does NOT currently allow creating managed types from Java.
-			sw.WriteLine ();
-			if (ctor.Annotations != null)
-				sw.WriteLine (ctor.Annotations);
-			sw.Write ("\tpublic ");
-			sw.Write (name);
-			sw.Write (" (");
-			sw.Write (ctor.Params);
-			sw.Write (')');
-			sw.WriteLine (ctor.ThrowsDeclaration);
-			sw.WriteLine ("\t{");
-			sw.Write ("\t\tsuper (");
-			sw.Write (ctor.SuperCall);
-			sw.WriteLine (");");
-#if MONODROID_TIMING
-			sw.WriteLine ("\t\tandroid.util.Log.i(\"MonoDroid-Timing\", \"{0}..ctor({1}): time: \"+java.lang.System.currentTimeMillis());", name, ctor.Params);
-#endif
-			if (!CannotRegisterInStaticConstructor (type)) {
-				sw.Write ("\t\tif (getClass () == ");
-				sw.Write (name);
-				sw.WriteLine (".class) {");
-				sw.Write ("\t\t\t");
-				switch (CodeGenerationTarget) {
-					case JavaPeerStyle.JavaInterop1:
-						sw.Write ("net.dot.jni.ManagedPeer.construct (this, \"");
-						sw.Write (ctor.JniSignature);
-						sw.Write ("\", new java.lang.Object[] { ");
-						sw.Write (ctor.ActivateCall);
-						sw.WriteLine (" });");
-						break;
-					default:
-						sw.Write ("mono.android.TypeManager.Activate (\"");
-						sw.Write (type.GetPartialAssemblyQualifiedName (cache));
-						sw.Write ("\", \"");
-						sw.Write (ctor.ManagedParameters);
-						sw.Write ("\", this, new java.lang.Object[] { ");
-						sw.Write (ctor.ActivateCall);
-						sw.WriteLine (" });");
-						break;
-				}
-				sw.WriteLine ("\t\t}");
-			}
-			sw.WriteLine ("\t}");
-		}
-
-		void GenerateApplicationConstructor (TextWriter sw)
-		{
-			if (!JavaNativeTypeManager.IsApplication (type, cache)) {
-				return;
-			}
-
-			sw.WriteLine ();
-			sw.Write ("\tpublic ");
-			sw.Write (name);
-			sw.WriteLine (" ()");
-			sw.WriteLine ("\t{");
-			sw.WriteLine ("\t\tmono.MonoPackageManager.setContext (this);");
-			sw.WriteLine ("\t}");
-		}
-
-		void GenerateExportedField (JavaFieldInfo field, TextWriter sw)
-		{
-			sw.WriteLine ();
-			if (field.Annotations != null)
-				sw.WriteLine (field.Annotations);
-			sw.Write ("\t");
-			sw.Write (field.GetJavaAccess ());
-			sw.Write (' ');
-			if (field.IsStatic)
-				sw.Write ("static ");
-			sw.Write (field.TypeName);
-			sw.Write (' ');
-			sw.Write (field.FieldName);
-			sw.Write (" = ");
-			sw.Write (field.InitializerName);
-			sw.WriteLine (" ();");
-		}
-
-		void GenerateMethod (Signature method, TextWriter sw)
-		{
-			sw.WriteLine ();
-			if (method.Annotations != null)
-				sw.WriteLine (method.Annotations);
-			sw.Write ("\t");
-			sw.Write (method.IsExport ? method.JavaAccess : "public");
-			sw.Write (' ');
-			if (method.IsStatic)
-				sw.Write ("static ");
-			sw.Write (method.Retval);
-			sw.Write (' ');
-			sw.Write (method.JavaName);
-			sw.Write (" (");
-			sw.Write (method.Params);
-			sw.Write (')');
-			sw.WriteLine (method.ThrowsDeclaration);
-			sw.WriteLine ("\t{");
-#if MONODROID_TIMING
-			sw.WriteLine ("\t\tandroid.util.Log.i(\"MonoDroid-Timing\", \"{0}.{1}: time: \"+java.lang.System.currentTimeMillis());", name, method.Name);
-#endif
-			sw.Write ("\t\t");
-			sw.Write (method.Retval == "void" ? String.Empty : "return ");
-			sw.Write ("n_");
-			sw.Write (method.Name);
-			sw.Write (" (");
-			sw.Write (method.ActivateCall);
-			sw.WriteLine (");");
-
-			sw.WriteLine ("\t}");
-			sw.WriteLine ();
-			sw.Write ("\tprivate ");
-			if (method.IsStatic)
-				sw.Write ("static ");
-			sw.Write ("native ");
-			sw.Write (method.Retval);
-			sw.Write (" n_");
-			sw.Write (method.Name);
-			sw.Write (" (");
-			sw.Write (method.Params);
-			sw.WriteLine (");");
-		}
-
-		void WriteApplicationOnCreate (TextWriter sw, Action<TextWriter> extra)
-		{
-			sw.WriteLine ();
-			sw.WriteLine ("\tpublic void onCreate ()");
-			sw.WriteLine ("\t{");
-			extra (sw);
-			sw.WriteLine ("\t}");
-		}
-
-		void WriteInstrumentationOnCreate (TextWriter sw, Action<TextWriter> extra)
-		{
-			sw.WriteLine ();
-			sw.WriteLine ("\tpublic void onCreate (android.os.Bundle arguments)");
-			sw.WriteLine ("\t{");
-
-#if MONODROID_TIMING
-			sw.WriteLine ("\t\tandroid.util.Log.i(\"MonoDroid-Timing\", \"{0}.onCreate(Bundle): time: \"+java.lang.System.currentTimeMillis());", name);
-			sw.WriteLine ();
-#endif
-
-			sw.WriteLine ("\t\tandroid.content.Context context = getContext ();");
-			sw.WriteLine ();
-
-			if (!string.IsNullOrEmpty (MonoRuntimeInitialization)) {
-				sw.WriteLine (MonoRuntimeInitialization);
-				sw.WriteLine ();
-			}
-
-			extra (sw);
-			sw.WriteLine ("\t}");
-		}
-
-		StreamWriter OpenStream (string outputPath)
-		{
-			string destination = GetDestinationPath (outputPath);
-			Directory.CreateDirectory (Path.GetDirectoryName (destination));
-			return new StreamWriter (new FileStream (destination, FileMode.Create, FileAccess.Write));
-		}
-
-		/// <summary>
-		/// Returns a destination file path based on the package name of this Java type
-		/// </summary>
-		public string GetDestinationPath (string outputPath)
-		{
-			var dir = package.Replace ('.', Path.DirectorySeparatorChar);
-			return Path.Combine (outputPath, dir, name + ".java");
 		}
 	}
 }
