+++ src/Java.Interop.Tools.JavaCallableWrappers/Java.Interop.Tools.JavaCallableWrappers/JavaPeerStyle.cs
@@ -1,0 +1,7 @@
+﻿namespace Java.Interop.Tools.JavaCallableWrappers
+{
+	public enum JavaPeerStyle {
+		XAJavaInterop1,
+		JavaInterop1,
+	}
+}
